+++ src/cjs/utils/algorithms/Metaballs.d.ts
@@ -10,6 +10,7 @@
  * @date    2024-03-01
  */
 import { Circle } from "../../Circle";
+import { XYCoords } from "../../interfaces";
 interface MetaballsOptions {
     metaRadiusAddon: number;
 }
@@ -22,7 +23,20 @@ export declare class Metaballs {
     constructor(inputCircles: Array<Circle>);
     rebuild(options: MetaballsOptions): void;
     private checkCircleFullyContained;
-    detectContainmentStatus(): void;
-    rebuildContainingCircles(options: MetaballsOptions): void;
+    private detectContainmentStatus;
+    private rebuildContainingCircles;
+    private detectInverseCircleHoles;
+    static metaballsUtils: {
+        detectHoles: (circles: Array<Circle>) => Array<number[]>;
+        /**
+         * Find a single hole group belonging to the circle at the given index.
+         * @param circles
+         * @param nonVisitedSet
+         * @param index
+         * @returns
+         */
+        detectHoleGroup: (circles: Array<Circle>, nonVisitedSet: Set<number>, index: number) => Array<number>;
+        anyCircleContainsPoint: (circles: Array<Circle>, point: XYCoords, ignoreCircleIndex: number) => boolean;
+    };
 }
 export {};