+++ src/cjs/utils/algorithms/Metaballs.js
@@ -74,9 +74,11 @@ var Metaballs = /** @class */ (function () {
                     circlePointsA: circlePointsA,
                     circlePointsB: circlePointsB
                 });
-            }
-        }
+            } // END for j
+        } // END for i
+        this.detectInverseCircleHoles();
     };
+    //   public getInnerPath;
     Metaballs.prototype.checkCircleFullyContained = function (circleIndex) {
         for (var i = 0; i < this.inputCircles.length; i++) {
             if (circleIndex == i) {
@@ -103,6 +105,73 @@ var Metaballs = /** @class */ (function () {
         for (var i = 0; i < this.circlesOfInterest.length; i++) {
             var containingCircle = new Circle_1.Circle(this.circlesOfInterest[i].center, this.circlesOfInterest[i].radius + options.metaRadiusAddon);
             this.containingCircles.push(containingCircle);
+        }
+    };
+    Metaballs.prototype.detectInverseCircleHoles = function () {
+        var inverseCircles = this.inverseCirclesPairs.reduce(function (accu, pair) {
+            accu.push(pair.inverseCircleA, pair.inverseCircleB);
+            return accu;
+        }, []);
+        console.log("inverseCircles", inverseCircles);
+        var holeGroupIndices = Metaballs.metaballsUtils.detectHoles(inverseCircles);
+        console.log("holeGroupIndices", holeGroupIndices);
+    };
+    Metaballs.metaballsUtils = {
+        // +---------------------------------------------------------------------------------
+        // | A hole can be found this way: a group of inverse circles with mutually contained centers.
+        // +-------------------------------
+        detectHoles: function (circles) {
+            var n = circles.length;
+            console.log("circles", circles);
+            // var isInverseCircleVisited = arrayFill(n, false); // Array<number>
+            var visitedCount = 0;
+            var nonVisitedSet = new Set();
+            var holeGroups = [];
+            for (var i = 0; i < n; i++) {
+                nonVisitedSet.add(i);
+            }
+            var iteration = 0; // A safety stop
+            while (visitedCount < n && iteration++ < n + 1) {
+                var curIndex = Array.from(nonVisitedSet)[Math.floor(Math.random() * nonVisitedSet.size)];
+                var holeGroupIndices = Metaballs.metaballsUtils.detectHoleGroup(circles, nonVisitedSet, curIndex);
+                visitedCount += holeGroupIndices.length;
+                holeGroups.push(holeGroupIndices);
+            }
+            return holeGroups;
+        },
+        /**
+         * Find a single hole group belonging to the circle at the given index.
+         * @param circles
+         * @param nonVisitedSet
+         * @param index
+         * @returns
+         */
+        detectHoleGroup: function (circles, nonVisitedSet, index) {
+            var holeGroupIndices = [index];
+            // Mark as visited
+            nonVisitedSet.delete(index);
+            for (var i = 0; i < circles.length; i++) {
+                if (!nonVisitedSet.has(i)) {
+                    // Already visited
+                    continue;
+                }
+                // Circles mutually contain their centers?
+                var circleA = circles[index];
+                var circleB = circles[i];
+                if (circleA.containsPoint(circleB.center) && circleB.containsPoint(circleA.center)) {
+                    holeGroupIndices.push(i);
+                    nonVisitedSet.delete(i);
+                }
+            }
+            return holeGroupIndices;
+        },
+        anyCircleContainsPoint: function (circles, point, ignoreCircleIndex) {
+            for (var i = 0; i < circles.length; i++) {
+                if (i != ignoreCircleIndex && circles[i].containsPoint(point)) {
+                    return true;
+                }
+            }
+            return false;
         }
     };
     return Metaballs;