+++ src/esm/utils/algorithms/Metaballs.js
@@ -71,9 +71,11 @@ export class Metaballs {
                     circlePointsA: circlePointsA,
                     circlePointsB: circlePointsB
                 });
-            }
-        }
+            } // END for j
+        } // END for i
+        this.detectInverseCircleHoles();
     }
+    //   public getInnerPath;
     checkCircleFullyContained(circleIndex) {
         for (var i = 0; i < this.inputCircles.length; i++) {
             if (circleIndex == i) {
@@ -102,5 +104,72 @@ export class Metaballs {
             this.containingCircles.push(containingCircle);
         }
     }
+    detectInverseCircleHoles() {
+        const inverseCircles = this.inverseCirclesPairs.reduce((accu, pair) => {
+            accu.push(pair.inverseCircleA, pair.inverseCircleB);
+            return accu;
+        }, []);
+        console.log("inverseCircles", inverseCircles);
+        var holeGroupIndices = Metaballs.metaballsUtils.detectHoles(inverseCircles);
+        console.log("holeGroupIndices", holeGroupIndices);
+    }
 }
+Metaballs.metaballsUtils = {
+    // +---------------------------------------------------------------------------------
+    // | A hole can be found this way: a group of inverse circles with mutually contained centers.
+    // +-------------------------------
+    detectHoles: (circles) => {
+        const n = circles.length;
+        console.log("circles", circles);
+        // var isInverseCircleVisited = arrayFill(n, false); // Array<number>
+        var visitedCount = 0;
+        const nonVisitedSet = new Set();
+        const holeGroups = [];
+        for (var i = 0; i < n; i++) {
+            nonVisitedSet.add(i);
+        }
+        var iteration = 0; // A safety stop
+        while (visitedCount < n && iteration++ < n + 1) {
+            const curIndex = Array.from(nonVisitedSet)[Math.floor(Math.random() * nonVisitedSet.size)];
+            const holeGroupIndices = Metaballs.metaballsUtils.detectHoleGroup(circles, nonVisitedSet, curIndex);
+            visitedCount += holeGroupIndices.length;
+            holeGroups.push(holeGroupIndices);
+        }
+        return holeGroups;
+    },
+    /**
+     * Find a single hole group belonging to the circle at the given index.
+     * @param circles
+     * @param nonVisitedSet
+     * @param index
+     * @returns
+     */
+    detectHoleGroup: (circles, nonVisitedSet, index) => {
+        const holeGroupIndices = [index];
+        // Mark as visited
+        nonVisitedSet.delete(index);
+        for (var i = 0; i < circles.length; i++) {
+            if (!nonVisitedSet.has(i)) {
+                // Already visited
+                continue;
+            }
+            // Circles mutually contain their centers?
+            const circleA = circles[index];
+            const circleB = circles[i];
+            if (circleA.containsPoint(circleB.center) && circleB.containsPoint(circleA.center)) {
+                holeGroupIndices.push(i);
+                nonVisitedSet.delete(i);
+            }
+        }
+        return holeGroupIndices;
+    },
+    anyCircleContainsPoint: (circles, point, ignoreCircleIndex) => {
+        for (var i = 0; i < circles.length; i++) {
+            if (i != ignoreCircleIndex && circles[i].containsPoint(point)) {
+                return true;
+            }
+        }
+        return false;
+    }
+};
 //# sourceMappingURL=Metaballs.js.map