+++ src/ts/CircleSector.ts
@@ -4,11 +4,13 @@
  * @modified 2021-01-20 Added UID.
  * @modified 2021-02-26 Fixed an error in the svg-arc-calculation (case angle<90deg and anti-clockwise).
  * @modified 2024-01-30 Added a missing type in the `describeSVGArc` function.
- * @version  1.1.2
+ * @modified 2024-03-01 Added the `getStartPoint` and `getEndPoint` methods.
+ * @version  1.2.0
  **/
 
 import { Circle } from "./Circle";
 import { UIDGenerator } from "./UIDGenerator";
+import { Vertex } from "./Vertex";
 import { SVGPathParams, SVGSerializable, UID, XYCoords } from "./interfaces";
 
 /**
@@ -82,9 +84,38 @@ export class CircleSector implements SVGSerializable {
   }
 
   /**
+   * Get the sectors starting point (on the underlying circle, located at the start angle).
+   *
+   * @method getStartPoint
+   * @instance
+   * @memberof CircleSector
+   * @return {Vertex} The sector's stating point.
+   */
+  getStartPoint(): Vertex {
+    return this.circle.vertAt(this.startAngle);
+  }
+
+  /**
+   * Get the sectors ending point (on the underlying circle, located at the end angle).
+   *
+   * @method getEndPoint
+   * @instance
+   * @memberof CircleSector
+   * @return {Vertex} The sector's ending point.
+   */
+  getEndPoint(): Vertex {
+    return this.circle.vertAt(this.endAngle);
+  }
+
+  /**
    * This function should invalidate any installed listeners and invalidate this object.
    * After calling this function the object might not hold valid data any more and
    * should not be used.
+   *
+   * @method destroy
+   * @instance
+   * @memberof CircleSector
+   * @return {void}
    */
   destroy() {
     this.circle.destroy();